+++ Sources/ColorPaletteCodable/coders/palette/PaintShopProPaletteCoder.swift
@@ -81,9 +81,9 @@ public extension PAL.Coder.PaintShopPro {
 					continue
 				}
 
-				let re = max(0, min(1, Float32(rv) / 255.0))
-				let ge = max(0, min(1, Float32(gv) / 255.0))
-				let be = max(0, min(1, Float32(bv) / 255.0))
+				let re = max(0, min(1, Double(rv) / 255.0))
+				let ge = max(0, min(1, Double(gv) / 255.0))
+				let be = max(0, min(1, Double(bv) / 255.0))
 				let c = PAL.Color(rf: re, gf: ge, bf: be)
 				palette.colors.append(c)
 			}
